// src/app/modules/ususarios/model/usuario-dto.ts
export interface UsuarioDTO {
	idUsuario: number;
	apellidoPaterno: string;
	apellidoMaterno: string;
	nombre: string;
	correo: string;
	folio: string;
	matricula: string;
	nombreUsuario: string;
	password: string;
	idEstatus: number;
	idPerfil: number;
	curp: string;
	desGenero: string;
	edad: number;
	fechaNacimiento: Date;
	genero: number;
	idAsentamiento: number;
	asentamiento: string;
	desFachada: string;
	calle: string;
	numeroExterior: string;
	numeroInterior: string;
	idDireccion: number;
	idMunicipio: number;
	municipio: string;
	idEstado: number;
	estado: string;
	colonia: string;
}